pyproject.toml
[tool.poetry]
name = "homebot"
version = "0.1.0"
description = "The smart bot to the rescue to automate tedious tasks around your home"
authors = ["Dennis Muth <d.muth@gmx.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
schiene = "^0.23"
slackclient = "^2.5.0"
terminaltables = "^3.1.0"
fire = "^0.2.1"
typeguard = "^2.7.0"
beautifulsoup4 = "^4.8.2"
pandas = "^0.25.3"
lxml = "^4.4.2"
httpx = "^0.9.5"
mako = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
pytest-cov = "^2.8.1"
flake8 = "^3.7.9"
pylint = "^2.4.4"
mypy = "^0.761"
invoke = "^1.3.0"
pytest-asyncio = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
